add footer to route layout, conditionally render it on item page
