+++ apps/site/components/client/index.ts
@@ -9,6 +9,7 @@ export * from './ChannelCard2'
 export * from './ChannelDialog'
 export * from './AddToChannelDialog'
 export * from './ChannelMembersPopover'
+export * from './Footer'
 export * from './Header'
 export * from './IndexInfoToggle'
 export * from './ItemDropdown'
